sssd: fix missing python3-core run-time dependency
The internal tool sss_analyze is a python script run by the sssctl
command analyze.

The script sss_analyze is shipped by the package sssd since 2.6.0.

However, the package sssd lacks installing the python interpreter that
is required to run the script.

This adds the missing run-time dependency python3-core to ensure the
interpreter python3 comes along the package sssd.

Fixes:

	root@qemux86-64:~# sssctl analyze
	env: can't execute 'python3': No such file or directory
	Command '/usr/libexec/sssd/sss_analyze' failed with [127]

Signed-off-by: Gaël PORTAY <gael.portay+rtone@gmail.com>